+++ samples/usage/simple-script/simple-script.ts
@@ -6,7 +6,7 @@ Note: [🔁] In your app you will be importing '@promptbook/core' instead of '..
 
 import colors from 'colors';
 import * as dotenv from 'dotenv';
-import { readFile, writeFile } from 'fs/promises';
+import { writeFile } from 'fs/promises';
 import { forTime } from 'waitasecond';
 import {
     createPipelineExecutor,
@@ -54,8 +54,8 @@ async function main() {
 
     const pipeline = await collection.getPipelineByUrl(
         //`https://promptbook.studio/samples/foreach-list.ptbk.md`,
-        `https://promptbook.studio/samples/foreach-csv.ptbk.md`,
-        //`https://promptbook.studio/samples/simple-knowledge.ptbk.md`,
+        //`https://promptbook.studio/samples/foreach-csv.ptbk.md`,
+        `https://promptbook.studio/samples/simple-knowledge.ptbk.md`,
         // `https://promptbook.studio/samples/simple.ptbk.md`,
         // `https://promptbook.studio/samples/language-capabilities.ptbk.md`,
     );
@@ -89,15 +89,15 @@ async function main() {
             Kate
         `),
         /**/
-        /**/
+        /*/
         // https://promptbook.studio/samples/foreach-csv.ptbk.md
         customers: await readFile('./samples/pipelines/85-foreach.csv', 'utf-8'),
         /**/
-        /*/
+        /**/
         // https://promptbook.studio/samples/simple-knowledge.ptbk.md
         eventTitle: 'LinkedIn',
-        eventDescription: 'Professional LinkedIn profile',
-        rules: 'Write best text for LinkedIn in first person',
+        eventDescription: 'Professional CV',
+        rules: 'Write best text for CV',
         /**/
     };
     const { isSuccessful, errors, warnings, outputParameters, executionReport, usage } = await pipelineExecutor(