+++ src/collection/constructors/createCollectionFromDirectory.test.ts
@@ -56,7 +56,6 @@ describe('createCollectionFromDirectory', () => {
         expect.assertions(1);
         const collection = await createCollectionFromDirectory('./samples/pipelines', {
             llmTools: null,
-            filesystemTools: null,
             isVerbose: true,
             isRecursive: false,
             isLazyLoaded: false,
@@ -76,7 +75,6 @@ describe('createCollectionFromDirectory', () => {
 
         const collection = await createCollectionFromDirectory('./samples/pipelines', {
             llmTools: null,
-            filesystemTools: null,
             isVerbose: true,
             isRecursive: false,
             isLazyLoaded: true,
@@ -96,7 +94,6 @@ describe('createCollectionFromDirectory', () => {
 
         const collection = await createCollectionFromDirectory('./samples/pipelines', {
             llmTools: null,
-            filesystemTools: null,
             isVerbose: true,
             isRecursive: false,
         });
@@ -115,7 +112,6 @@ describe('createCollectionFromDirectory', () => {
             (async () => {
                 const collection = await createCollectionFromDirectory('./samples/pipelines', {
                     llmTools: null,
-                    filesystemTools: null,
                     isVerbose: true,
                     // Note: Including subdirectories BUT lazy-loaded so it should not crash even if there are errors
                     isRecursive: true,
@@ -130,7 +126,6 @@ describe('createCollectionFromDirectory', () => {
             (async () => {
                 const collection = await createCollectionFromDirectory('./samples/pipelines', {
                     llmTools: null,
-                    filesystemTools: null,
                     isVerbose: true,
                     // Note: Including subdirectories BUT lazy-loaded so it should not crash even if there are errors
                     isRecursive: true,